Added list/item not found check to remove_list_item.

+++ web/NotebookView.pl
@@ -631,13 +631,14 @@ sub remove_list_item {
 	#return "Permission denied" unless $USER->is_admin || $USER->is_owner( dsg => $dsgid );
 	
 	my $list = $coge->resultset('List')->find($lid);
+	return 0 unless $list;
 	return 0 if ($list->locked && !$USER->is_admin);
 	
 	my $item_type = $opts{item_type};
 	my $item_id = $opts{item_id};
 	
 	my $lc = $coge->resultset('ListConnector')->find( { parent_id => $lid, child_id => $item_id, child_type =>$item_type } );
-	$lc->delete();
+	$lc->delete() if $lc;
 	
 	return 1;
 }
